Add files via upload
Hybrid Model Approach: 
The present Notebook will be composed by the best model from the evaluation of SARIMAX. Next, the LSTM will come into play to attempt to predict the residuals  of the SARIMAX model. 
Like the stand alone models, the LSTM architecture requires that the residuals are scaled. Therefore, different types of scaling and sequence length will be performed to check which one will yield better results using the residuals as target. 

Next step, Evaluation table and Justification to select the best model for each sequence length and consequently the ultimate model for the hybrid approach.